+++ org.titou10.jtb.build/CHANGELOG.md
@@ -1,13 +1,22 @@
-2024-??-?? v6.7.0
+2024-??-?? v6.8.0
 ------------------
 - Important Changes
-    -
+    - 
 - New features:
-    -
+    - Artemis 2.x: Added keystore* properties
 - Bugs fixed:
-    -
+    - Fixes #182: JTB fails to "pretty print" JSON arrays
+    - Artemis 2.x: categorize MULTICAST+ANYCAST addresses as queues and not topics
+- Other:  
+    - Replaced javax/jaxb by jakarta/jaxb
+    
+2024-09-12 v6.7.0
+------------------
+- New features:
+    - Artemis 2.x: added "trustStoreType" property
 - Other:  
     - Upgraded Apache Artemis 2.x jars to v2.37.0
+    - Upgraded eclipse RCP to v2024-09 (v4.33)
 
 2024-06-20 v6.6.0
 ------------------